Fixes validation of maxIdleTimeout in metastore http server

+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/HiveMetaStore.java
@@ -465,7 +465,7 @@ public class HiveMetaStore extends ThriftHiveMetastore {
     connector.setPort(port);
     connector.setReuseAddress(true);
     // TODO: What should the idle timeout be for the metastore? Currently it is 30 minutes
-    int maxIdleTimeout = (int)MetastoreConf.getTimeVar(conf, ConfVars.METASTORE_THRIFT_HTTP_MAX_IDLE_TIME,
+    long maxIdleTimeout = MetastoreConf.getTimeVar(conf, ConfVars.METASTORE_THRIFT_HTTP_MAX_IDLE_TIME,
         TimeUnit.MILLISECONDS);
     connector.setIdleTimeout(maxIdleTimeout);
     // TODO: AcceptQueueSize needs to be higher for HMS
